create template file for build/__main__.py construction
