+++ starfyre/__main__.py
@@ -1,6 +1,7 @@
 from starfyre import compile
 from pathlib import Path
 from starfyre.file_router import FileRouter
+from starfyre.main_template import get_main_template
 
 import sys
 import os
@@ -16,80 +17,35 @@ def write_js_file(path):
     shutil.copy(str(js_store), path + "/dist/store.js")
 
 def build_routes_output(generated_routes, path):
+    """
+    Build the HTML output files for each generated route and create the main execution file.
+
+    Parameters:
+    - generated_routes (list): List of generated route names.
+    - path (str): Path to the project directory.
+
+    This function gets the HTML output for each generated route based on the provided `generated_routes` list.
+    It also creates the main execution file (__main__.py) that orchestrates rendering and writing HTML content to files.
+    The rendered HTML content is generated by importing and utilizing the corresponding components for each route.
+
+    Args:
+        generated_routes (list): A list of route names that have been generated.
+        path (str): The path to the project directory.
+    """
     write_js_file(path)
 
     output_file_path = path + "/build/__main__.py"
     init_file_path = path + "/build/__init__.py"
     user_routes = generated_routes[:]
     user_routes.insert(0, 'app')
-    
+    main_script_content = get_main_template(user_routes=user_routes, path=path)
+
     # create empty __init__.py file
     with open(init_file_path, "w"):
         pass
 
     with open(output_file_path, "w") as f:
-        f.write(
-            f'''from starfyre import create_component, render_root
-from starfyre.exceptions import IndexFileConflictError
-
-import os
-import sys
-from pathlib import Path
-import importlib
-
-def generate_pages(generated_routes, path):
-    """
-    Generate HTML pages for each route in the provided list of routes.
-
-    Parameters:
-    - generated_routes (list): List of generated routes.
-    - path (str): Path to the project directory.
-
-    This function generates HTML pages for each route provided in the `generated_routes` list.
-    It imports the necessary components for each route, renders them using Starfyre,
-    and writes the rendered content to corresponding HTML files.
-    """
-
-    out_dir = Path(path + "/dist").resolve()
-    root = Path(out_dir / "..").resolve()
-
-    # get the user defined project name
-    app_name = (str(root).split('/'))[-1]
-
-    for route_name in generated_routes:
-        print(f'route name is = {{route_name}}')
-        
-        if route_name.lower() == 'app':
-            component_key = 'app'  # For the 'app' component in `build/pages/__init__.py`
-        else:
-            component_key = route_name
-        
-        if route_name == 'app':
-            module_name = f"{{Path(app_name)}}.build.pages"
-        else:
-            module_name= f"{{Path(app_name)}}.build.pages.{{route_name}}"
-        
-        try:
-            module = importlib.import_module(module_name)
-            if component_key == 'app':
-                component = getattr(module, component_key, f"{{component_key}} does not exist")
-            else:
-                component = getattr(module, f'rendered_{{component_key}}')
-            result = str(component)
-        except ModuleNotFoundError:
-            print(f"Error: Could not import module '{{module_name}}'.")
-            continue
-
-        # write to component file
-        if route_name == 'app':
-            route_name = 'index'  # rename to index
-        with open(out_dir / f"{{route_name}}.html", "w") as html_file:
-            html_file.write("<script src='store.js'></script>")
-            html_file.write(result)
-
-if __name__ == '__main__':
-    generate_pages(generated_routes={user_routes}, path="{path}")
-    ''')
+        f.write(main_script_content)
 
 @click.command()
 @click.option("--path", help="Path to the project")
@@ -132,11 +88,9 @@ def main(path, build):
         result = subprocess.run(
             [sys.executable, "-m", "build"],
             cwd=path,
-            # stdout=subprocess.PIPE,
+            stdout=subprocess.PIPE,
             stderr=None,
         )
-
-        print(result.stdout.decode("utf-8"))
 
 
 if __name__ == "__main__":
